Super basic solver.
One key thing this solver needs is tracking to prevent loops.

The simplest form is to avoid ever making a move that is the inverse of
the previous move, but it may be possible to just always hash the state
and immediately undo any other move that matches a state in the current
stack of states.